% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.r
\name{alert_from_template}
\alias{alert_from_template}
\title{Alert From a Template}
\usage{
alert_from_template(varlist, template.file, figure = NULL)
}
\arguments{
\item{varlist}{a named list of variables that are to
inserted into the email content. Variable names must
exactly match the placeholders in the YAML template.
Required variables are \code{Station}, \code{Datetime}, \code{Stage},
and \code{Threshold}.}

\item{template.file}{The YAML template for the alert.}

\item{figure}{A \code{ggplot} object.}
}
\value{
A named list.
}
\description{
Construct the contents of an email/text alert from a YAML
template.
}
\keyword{internal}
