% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholds.r
\name{check_limit}
\alias{check_limit}
\title{Limit Check}
\usage{
check_limit(x, limit)
}
\arguments{
\item{x}{A vector of values.}

\item{limit}{The threshold to compare to.}
}
\value{
Logical \code{TRUE} if threshold is exceeded, \code{FALSE} otherwise.
}
\description{
Compare a variable against a maximum limit.
}
