% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholds.r
\name{check_rate}
\alias{check_rate}
\title{Rate Check}
\usage{
check_rate(x, tm, limit)
}
\arguments{
\item{x}{A vector of values.}

\item{tm}{A vector of timestamps.}

\item{limit}{The threshold to compare to.}
}
\value{
Logical \code{TRUE} if threshold is exceeded, \code{FALSE} otherwise.
}
\description{
Compare the rate of change of a variable against a threshold.
}
