% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.r
\name{check_rrds}
\alias{check_rrds}
\title{Check the Roaring River Distribution System}
\usage{
check_rrds(
  stage.threshold = 4.6,
  rate.threshold = 3.6e-05,
  outside.stage.threshold = 7.2,
  surge.threshold = 8,
  por = "6 hours",
  alert = TRUE,
  verbose = FALSE,
  stage.template = default.stage.template,
  rate.template = default.rate.template,
  surge.template = default.surge.template
)
}
\arguments{
\item{stage.threshold}{The threshold stage to trigger an alert.
default is 4.6 ft NAVD88.}

\item{rate.threshold}{The threshold rate to trigger an alert,
in units of ft/s. Default is 0.3 ft in 15 minutes.}

\item{outside.stage.threshold}{The threshold stage outside of the
system required to trigger a rate-based alert. default is
7.2 ft NAVD88.}

\item{surge.threshold}{The threshold stage prediction at Rio Vista
Bridge required to trigger a storm surge alert. default is 8.0 ft
NAVD88.}

\item{por}{The period of record to use for check. Default is to use
previous 6 hours of data.}

\item{alert}{If \code{TRUE}, email alerts will be sent if thresholds
are triggered.}

\item{verbose}{If \code{TRUE}, display messages on console.}

\item{stage.template}{The YAML template to use for the stage alerts.}

\item{rate.template}{The YAML template to use for the rate alerts.}

\item{surge.template}{The YAML template to use for the storm surge
alerts.}
}
\value{
\code{TRUE} if message is sent, \code{FALSE} if thresholds are not
exceeded.
}
\description{
Check stages in the Roaring River Distribution System
recorded on CDEC and send an email alert if stages exceed
the specified threshold.
}
