% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.r
\name{get_cnrfc_forecast}
\alias{get_cnrfc_forecast}
\title{Get CNRFC Forecast}
\usage{
get_cnrfc_forecast(location = "RVBC1")
}
\arguments{
\item{location}{The location for which to retrieve the forecast.
Default is Rio Vista "RVBC1".}
}
\description{
Get the CNRFC Forecast for a location.
}
