% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.r
\name{get_station_data}
\alias{get_station_data}
\title{Get Station Data}
\usage{
get_station_data(stations, sensors, por, prog, db)
}
\arguments{
\item{stations}{The CDEC station IDs to query.}

\item{sensors}{The CDEC sensor numbers to query.}
}
\value{
A dataframe.
}
\description{
Get station data from CDEC.
}
\keyword{internal}
