% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.r
\name{loop_rrds}
\alias{loop_rrds}
\title{Loop RRDS Messaging}
\usage{
loop_rrds(
  ...,
  check.interval = "1 hour",
  alert.interval = "1 day",
  error.template = default.error.template
)
}
\arguments{
\item{...}{Arguments to pass to \code{check_rrds}}

\item{check.interval}{Interval of time to wait between
consecutive checks.}

\item{alert.interval}{Interval of time to wait between
consecutive alerts.}

\item{error.template}{The YAML template to use for
constructing notifications of alert system failure.}
}
\description{
Run \code{check_rrds} in a perpetual loop.
}
