% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.r
\name{make_figure}
\alias{make_figure}
\title{Make Figure
Make a plot of RRDS stages from CDEC data.}
\usage{
make_figure(d, threshold)
}
\arguments{
\item{d}{The dataset construct the plot from.}

\item{threshold}{The threshold stage, plotted as
as a horizontal line.}
}
\value{
a \code{ggplot} object.
}
\description{
Make Figure
Make a plot of RRDS stages from CDEC data.
}
\keyword{internal}
