% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.r
\name{send_alert}
\alias{send_alert}
\title{Send Email Alert}
\usage{
send_alert(template, use_ssl = FALSE, verbose = FALSE)
}
\arguments{
\item{use_ssl, }{verbose additional arguments to \code{curl::send_mail()}.}

\item{Template}{a named list, i.e. output of \code{alert_from_template()}.}
}
\description{
Send an email alert.
}
\keyword{internal}
