% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.r
\name{use_template}
\alias{use_template}
\title{Use Alert Template}
\usage{
use_template(template.path)
}
\arguments{
\item{template.path}{The path to the template YAML file.}
}
\description{
Use the specified alert template for messaging.
}
